@@ -1,18 +1,19 @@
 package carcrowdsystem.ccs.entitys;
 
-import javax.persistence.*;
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+
 import java.util.List;
 
 @Entity
+@Table(name = "estacionamento")
 //@SequenceGenerator(name = "ID_INIT_100", sequenceName = "FUNCIONARIO_SEQ", initialValue = 100)
 // generator = "ID_INIT_100"
 public class EstacionamentoEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id_estacionamento")
     private Integer id;
 
     @NotBlank
@@ -20,10 +21,12 @@ public class EstacionamentoEntity {
     private String cnpj;
     @NotBlank
     @Size(min = 3)
+    @Column(name = "nome_estacionamento")
     private String nomeEstacionamento;
     @NotBlank
     private String cep;
     @NotBlank
+    @Column(name = "numero_endereco")
     private String numeroEndereco;
     private String telefone;
     @OneToMany
