@@ -1,17 +1,25 @@
 package carcrowdsystem.ccs.entitys;
 
-import javax.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+
+import jakarta.persistence.*;
 
 @Entity
+@Table(name = "valor_estacionamento")
 public class ValorEstacionamentoEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id_preco")
     private Integer id;
-    @ManyToOne
-    private EstacionamentoEntity estacionamento;
+    @Column(name = "primeira_hora")
     private Double primeiraHora;
+    @Column(name = "hora_adicional")
     private Double horaAdicional;
     private Double diaria;
+    @ManyToOne
+    @JoinColumn(name = "fk_estacionamento")
+    private EstacionamentoEntity estacionamento;
 
     public Integer getId() {
         return id;
