@@ -39,11 +39,11 @@ public class SecurityConfiguracao {
             new AntPathRequestMatcher("/webjars/**"),
             new AntPathRequestMatcher("/actuator/*"),
             new AntPathRequestMatcher("/arquivos/**"),
+            new AntPathRequestMatcher("/swagger-ui**"),
             new AntPathRequestMatcher("/swagger-ui/**"),
             new AntPathRequestMatcher("/h2-console/**"),
             new AntPathRequestMatcher("/api/public/**"),
             new AntPathRequestMatcher("/v3/api-docs/**"),
-            new AntPathRequestMatcher("/swagger-ui.html"),
             new AntPathRequestMatcher("/configuration/ui"),
             new AntPathRequestMatcher("/swagger-resources"),
             new AntPathRequestMatcher("/swagger-resources/**"),
@@ -51,6 +51,7 @@ public class SecurityConfiguracao {
             new AntPathRequestMatcher("/configuration/security"),
             new AntPathRequestMatcher("/api/public/authenticate"),
             new AntPathRequestMatcher("/ccs-dev/funcionarios/**"),
+            new AntPathRequestMatcher("/ccs-dev/estacionamentos/**"),
             new AntPathRequestMatcher("/ccs-dev/checkin/mensagem/**")
     };
 
